+++ backend/src/questions/views.py
@@ -1,5 +1,6 @@
 import datetime
 import itertools as it
+from starlette.authentication import requires
 from starlette.responses import (
     UJSONResponse,
     RedirectResponse,
@@ -18,44 +19,9 @@ from accounts.models import User
 
 
 async def questions_all(request):
-    '''
-    ### i don't now how to serialize fk relations with
-    .prefetch.related()
-    using .values() is ok but can get m2m relations
-    like GROUP BY raw SQL
-
-    questions = await Question.all().values(
-        'id',
-        'title',
-        'content',
-        'created',
-        'view',
-        'question_like',
-        'answer_count',
-        'accepted_answer',
-        'user__username',
-        'tags__name'# problem
-    )
-
-    ### than i use raw sql with transaction
-    but load test with Bombardier
-    give pymysql.err.OperationalError: (2013,
-    'Lost connection to MySQL server during query')
-
-    async with in_transaction() as conn:
-        questions = await conn.execute_query(
-            "SELECT question.*, user.username, GROUP_CONCAT(tag.name) as tags \
-            FROM question JOIN user ON user.id=question.user_id \
-            JOIN question_tag ON question.id = question_tag.question_id \
-            JOIN tag ON question_tag.tag_id = tag.id \
-            GROUP BY question.id ORDER BY question.id DESC"
-        )
-
-    ### SQLAlchemy doesn't have this problem
-
-    ### ended up using a Marshmallow for serilization
-    and it turned out just fine
-    '''
+    """
+    Marshmallow for serilization
+    """
     results = await Question.all().prefetch_related(
         "user", "tags").order_by("-id")
     # Serialize the queryset
@@ -118,10 +84,12 @@ async def question(request):
     """
     id = request.path_params["id"]
     result = await Question.get(id=id).prefetch_related("user", "tags")
-    answers_result = (await Answer.all()
-                      .prefetch_related("ans_user")
-                      .filter(question__id=id)
-                      .order_by("-id"))
+    answers_result = (
+        await Answer.all()
+        .prefetch_related("ans_user")
+        .filter(question__id=id)
+        .order_by("-id")
+    )
     # Serialize the queryset
     question = question_schema.dump(result)
     answers = answers_schema.dump(answers_result)
@@ -137,17 +105,21 @@ async def question(request):
     )
 
 
+@requires("authenticated")
 async def question_like(request):
     id = request.path_params["id"]
     results = await Question.get(id=id)
-    # update question likes and decrease question views
-    # to avoid duplication of question views
-    results.question_like += 1
-    results.view -= 1
-    await results.save()
-    return RedirectResponse(
-        url="/questions", status_code=302
-    )
+    if request.user.is_authenticated:
+        # update question likes and decrease question views
+        # to avoid duplication of question views
+        results.question_like += 1
+        results.view -= 1
+        await results.save()
+        return RedirectResponse(
+            url="/", status_code=303
+        )
+    else:
+        return Response(status_code=403)
 
 
 async def question_create(request):
@@ -182,36 +154,40 @@ async def question_create(request):
                 await tag.save()
                 tags.append(tag)
                 await query.tags.add(tags[idx])
-            return RedirectResponse(url="/questions", status_code=302)
+            return RedirectResponse(url="/questions", status_code=303)
         return Response(
             "Tags must be comma-separated",
             status_code=422
         )
 
 
+@requires("authenticated")
 async def question_edit(request):
     """
     Question edit form
     """
     id = request.path_params["id"]
     session_user = request.user.username
-    results = await User.get(username=session_user)
-    question = await Question.get(id=id)
-    form = await request.json()
-    title = form["title"]
-    content = form["content"]
-    await Question.filter(id=id).update(
-        title=title,
-        slug="-".join(title.lower().split()),
-        content=content,
-        created=question.created,
-        view=question.view,
-        question_like=question.question_like,
-        answer_count=question.answer_count,
-        accepted_answer=question.accepted_answer,
-        user_id=results.id,
-    )
-    return RedirectResponse(url="/questions/", status_code=302)
+    question = await Question.get(id=id).prefetch_related('user')
+    if request.method == "PUT" and question.user.username == session_user:
+        form = await request.json()
+        title = form["title"]
+        content = form["content"]
+        await Question.filter(id=id).update(
+            title=title,
+            slug="-".join(title.lower().split()),
+            content=content,
+            created=question.created,
+            view=question.view,
+            question_like=question.question_like,
+            answer_count=question.answer_count,
+            accepted_answer=question.accepted_answer,
+            user_id=question.user.id,
+        )
+        # 303 status code for redirect after update
+        return RedirectResponse(url="/questions/", status_code=303)
+    else:
+        return Response(status_code=403)
 
 
 async def questions_user(request):
@@ -229,21 +205,27 @@ async def questions_user(request):
     )
 
 
+@requires("authenticated")
 async def question_delete(request):
     """
     Delete question
     """
     id = request.path_params["id"]
-    async with in_transaction() as conn:
-        await conn.execute_query(
-            f"DELETE FROM tag WHERE tag.id IN \
+    session_user = request.user.username
+    results = await Question.get(id=id).prefetch_related('user')
+    if request.method == "DELETE" and results.user.username == session_user:
+        async with in_transaction() as conn:
+            await conn.execute_query(
+                f"DELETE FROM tag WHERE tag.id IN \
                 (SELECT question_tag.tag_id FROM question \
-                JOIN question_tag ON question_tag.question_id = question.id \
+                JOIN question_tag \
+                ON question_tag.question_id = question.id \
                 WHERE question.id={id})"
-        )
-    await Question.get(id=id).delete()
-    response = RedirectResponse(url="/", status_code=302)
-    return response
+            )
+        await Question.get(id=id).delete()
+        # 303 status code for redirect after delete
+        response = RedirectResponse(url="/", status_code=303)
+        return response
 
 
 async def answer_create(request):
@@ -269,23 +251,27 @@ async def answer_create(request):
         results.answer_count += 1
         await results.save()
         return RedirectResponse(
-            url="/questions/", status_code=302
+            url="/questions/", status_code=303
         )
 
 
+@requires("authenticated")
 async def answer_like(request):
     id = request.path_params["id"]
     result = await Answer.get(id=id)
     question = await Question.get(id=result.question_id)
-    # update answer likes and decrease question views
-    # to avoid duplication of question views
-    result.answer_like += 1
-    await result.save()
-    question.view -= 1
-    await question.save()
-    return RedirectResponse(
-        url="/questions/", status_code=302
-    )
+    if request.user.is_authenticated:
+        # update answer likes and decrease question views
+        # to avoid duplication of question views
+        result.answer_like += 1
+        await result.save()
+        question.view -= 1
+        await question.save()
+        return RedirectResponse(
+            url="/questions/", status_code=303
+        )
+    else:
+        return Response(status_code=403)
 
 
 async def answer_accept(request):
@@ -297,7 +283,7 @@ async def answer_accept(request):
     question.accepted_answer = 1
     await question.save()
     return RedirectResponse(
-        url="/questions/", status_code=302
+        url="/questions/", status_code=303
     )
 
 
@@ -314,40 +300,54 @@ async def answers_user(request):
     )
 
 
+@requires("authenticated")
 async def answer_edit(request):
     """
     Answer edit form
     """
     id = request.path_params["id"]
-    answer = await Answer.get(id=id)
-    form = await request.json()
-    content = form["content"]
-    await Answer.filter(id=id).update(
-        content=content,
-        created=answer.created,
-        answer_like=answer.answer_like,
-        is_accepted_answer=answer.is_accepted_answer,
-        question_id=answer.question_id,
-        ans_user_id=answer.ans_user_id,
-    )
-    return RedirectResponse(url="/questions/", status_code=302)
+    session_user = request.user.username
+    answer = await Answer.get(id=id).prefetch_related('ans_user')
+    if request.method == "PUT" and answer.ans_user.username == session_user:
+        form = await request.json()
+        content = form["content"]
+        await Answer.filter(id=id).update(
+            content=content,
+            created=answer.created,
+            answer_like=answer.answer_like,
+            is_accepted_answer=answer.is_accepted_answer,
+            question_id=answer.question_id,
+            ans_user_id=answer.ans_user_id,
+        )
+        # 303 status code for redirect after update
+        return RedirectResponse(url="/questions/", status_code=303)
+    else:
+        return Response(status_code=403)
 
 
+@requires("authenticated")
 async def answer_delete(request):
     """
     Delete answer
     """
     id = request.path_params["id"]
-    answer = await Answer.get(id=id)
+    session_user = request.user.username
+    answer = (
+        await Answer.get(id=id)
+        .prefetch_related('ans_user')
+        .filter(ans_user__username=session_user)
+    )
     results = await Question.get(id=answer.question_id)
-    # decrease question answer count
-    results.answer_count -= 1
-    if answer.is_accepted_answer:
-        results.accepted_answer = False
-    await results.save()
-    await Answer.get(id=id).delete()
-    response = RedirectResponse(url="/", status_code=302)
-    return response
+    if request.method == "DELETE" and answer.ans_user.username == session_user:
+        # decrease question answer count
+        results.answer_count -= 1
+        if answer.is_accepted_answer:
+            results.accepted_answer = False
+        await results.save()
+        await Answer.get(id=id).delete()
+        # 303 status code for redirect after delete
+        response = RedirectResponse(url="/", status_code=303)
+        return response
 
 
 async def tags(request):
